Fixed Bugs Before Submitting Project
Fixed a bug where the app would crash if you tried to
save a reminder without keying anything into the
EditText fields.

Fixed a bug where the map would not zoom to the location
of the reminder when you clicked on the notification.

Fixed a bug where the logout menu item didn't work.